+++ tests/unit/test_spice_utils.py
@@ -1,3 +1,3 @@
 version https://git-lfs.github.com/spec/v1
-oid sha256:a4a618f9f47ca9e84fdb996cfbc2219a710bbd7935263bad49cc58e38d876a30
-size 7364
+oid sha256:84f4806a0e4e33e112f1dd3b94d2b6981151264bb0f9ba5efe7f66eb889e2f21
+size 7339