+++ tests/integration/test_kernel_maker.py
@@ -1,3 +1,3 @@
 version https://git-lfs.github.com/spec/v1
-oid sha256:18d4aa84b41e0240d0434ff91eece0583b09b0c64cf54c61dc4756fae85f155f
-size 13972
+oid sha256:55c861af599dcd21e8e6f18d99f9ef938e4dd22d704a6976613e7b4583838cc2
+size 13552