+++ tests/plugins/data_path_fixtures.py
@@ -1,3 +1,3 @@
 version https://git-lfs.github.com/spec/v1
-oid sha256:de4ecdb7c30329c9130bdc6ee51f1707875c03d9d2180457e0748e531c37d205
-size 3375
+oid sha256:f7d1b4704fcb1b28e0a68f277ffb64e5335c4801e9328a2356ccc18e63304fcd
+size 3382