+++ tests/unit/test_io/test_manifest.py
@@ -1,3 +1,3 @@
 version https://git-lfs.github.com/spec/v1
-oid sha256:a195f21d26bde6aabe0843141e05d14ea925e131b7d5e23f6b6f2a95ff49821e
-size 4317
+oid sha256:c62d4a6dcdcb74f35f2d731c69c2b620fb22a6c3162b846babcb130554360abe
+size 4402